renaming sparse_stoch_mat module
