+++ src/flowstab/sparse_stoch_mat.py
